Code is mostly and or completley cleaned up, needs bugfix
The geometry flickers. I tried transposing the matrices, I tired not transposing them, I tried transposing some of them, everything. Flickering could be caused by transposing the projection matrix in every frame, but this doesnt seem to be the problem for me since I NEVER transpose either the view or projection matrices at it is currenty, but the flickering still appears. As if magic though, everything else seems correct even though transpose is only done on the objects world matrices. In other words, it needs bugfixing.